refactor: remove the embedded subscription component
